requirements.txt
fastapi
uvicorn[standard]
chromadb
sentence-transformers
google-generativeai
python-dotenv
pydantic
